Switch yet again to FP16

+++ train_gpt2.py
@@ -174,7 +174,7 @@ for step in step_it:
             for _ in range(val_loss_steps):
                 x, y = val_loader.next_batch()
                 x, y = x.to(device), y.to(device)
-                with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+                with torch.autocast(device_type=device_type, dtype=torch.float16):
                     logits, loss = model(x, y)
                 loss = loss / val_loss_steps
                 val_loss_accum += loss.detach()
@@ -212,7 +212,7 @@ for step in step_it:
             mask = mask.to(device)
             # Get the logits
             with torch.no_grad():
-                with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+                with torch.autocast(device_type=device_type, dtype=torch.float16):
                     logits, loss = model(tokens)
                 pred_norm = get_most_likely_row(tokens, mask, logits)
             num_total += 1
@@ -245,7 +245,7 @@ for step in step_it:
         while xgen.size(1) < max_length:
             # Forward the model to get the logits
             with torch.no_grad():
-                with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+                with torch.autocast(device_type=device_type, dtype=torch.float16):
                     logits, loss = model(xgen) # (B, T, vocab_size)
                 # Take the logits at the last position
                 logits = logits[:, -1, :] # (B, vocab_size)
@@ -274,10 +274,10 @@ for step in step_it:
     for micro_step in range(grad_accum_steps):
         x, y = train_loader.next_batch()
         x, y = x.to(device), y.to(device)
-        # Added after video, this field is also used by the forward pass.
+        # Added after video, this field is also used by the forward pass
         if ddp:
             model.require_backward_grad_sync = (micro_step == grad_accum_steps - 1)
-        with torch.autocast(device_type=device_type, dtype=torch.bfloat16):
+        with torch.autocast(device_type=device_type, dtype=torch.float16):
             logits, loss = model(x, y)
         # We have to scale the loss to account for gradient accumulation,
         # because the gradients just add on each successive backward().
